Transferred the installation code to setup.py
